feat: set ID and description for VIP 411 and 412
