+++ simulations/vip-411/bscmainnet.ts
@@ -1,10 +1,10 @@
 import { expectEvents } from "../../src/utils";
 import { forking, testVip } from "../../src/vip-framework";
-import vip408 from "../../vips/vip-408/bscmainnet";
+import vip411 from "../../vips/vip-411/bscmainnet";
 import OMNICHAIN_PROPOSAL_SENDER_ABI from "./abi/OmnichainProposalSender.json";
 
 forking(45031950, async () => {
-  testVip("Ethena", await vip408(), {
+  testVip("Ethena", await vip411(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(
         txResponse,
