+++ simulations/vip-411/bsctestnet.ts
@@ -1,10 +1,10 @@
 import { expectEvents } from "../../src/utils";
 import { forking, testVip } from "../../src/vip-framework";
-import vip408 from "../../vips/vip-408/bsctestnet";
+import vip411 from "../../vips/vip-411/bsctestnet";
 import OMNICHAIN_PROPOSAL_SENDER_ABI from "./abi/OmnichainProposalSender.json";
 
 forking(46452753, async () => {
-  testVip("Ethena", await vip408(), {
+  testVip("Ethena", await vip411(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(
         txResponse,
