+++ simulations/vip-412/bscmainnet.ts
@@ -1,10 +1,10 @@
 import { expectEvents } from "../../src/utils";
 import { forking, testVip } from "../../src/vip-framework";
-import vip407 from "../../vips/vip-407/bscmainnet";
+import vip412 from "../../vips/vip-412/bscmainnet";
 import OMNICHAIN_PROPOSAL_SENDER_ABI from "./abi/OmnichainProposalSender.json";
 
 forking(45031950, async () => {
-  testVip("Ethena", await vip407(), {
+  testVip("Ethena", await vip412(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(
         txResponse,
