+++ simulations/vip-412/bsctestnet-addendum.ts
@@ -1,10 +1,10 @@
 import { expectEvents } from "../../src/utils";
 import { forking, testVip } from "../../src/vip-framework";
-import vip409 from "../../vips/vip-409/bsctestnet";
+import vip412 from "../../vips/vip-412/bsctestnet-addendum";
 import OMNICHAIN_PROPOSAL_SENDER_ABI from "./abi/OmnichainProposalSender.json";
 
 forking(46452753, async () => {
-  testVip("Ethena", await vip409(), {
+  testVip("Ethena", await vip412(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(
         txResponse,
