+++ simulations/vip-412/ethereum.ts
@@ -4,22 +4,25 @@ import { Contract } from "ethers";
 import { parseUnits } from "ethers/lib/utils";
 import { ethers } from "hardhat";
 import { NETWORK_ADDRESSES } from "src/networkAddresses";
+import { setMaxStalePeriodInChainlinkOracle } from "src/utils";
 import { forking, pretendExecutingVip, testForkedNetworkVipCommands } from "src/vip-framework";
+import { checkIsolatedPoolsComptrollers } from "src/vip-framework/checks/checkIsolatedPoolsComptrollers";
 import { checkVToken } from "src/vip-framework/checks/checkVToken";
 import { checkInterestRate } from "src/vip-framework/checks/interestRateModel";
 
 import vip072 from "../../multisig/proposals/ethereum/vip-072/index";
+import { COMPTROLLER_ETHENA, USDC, sUSDe, vip411 } from "../../vips/vip-411/bscmainnet";
 import {
-  COMPTROLLER_ETHENA,
-  PRIME,
-  PT_USDe_27MAR2025,
-  PT_sUSDE_27MAR2025,
-  VPT_USDe_27MAR2025_ETHENA,
-  VPT_sUSDE_27MAR2025_ETHENA,
+  CONVERSION_INCENTIVE,
   VTOKEN_RECEIVER,
-  vip407,
-} from "../../vips/vip-407/bscmainnet";
+  VUSDC_Ethena,
+  VsUSDe_Ethena,
+  converterBaseAssets,
+  underlyingAddress,
+  vip412,
+} from "../../vips/vip-412/bscmainnet";
 import RESILIENT_ORACLE_ABI from "./abi/ResilientOracle.json";
+import SINGLE_TOKEN_CONVERTER_ABI from "./abi/SingleTokenConverter.json";
 import COMPTROLLER_ABI from "./abi/comptroller.json";
 import ERC20_ABI from "./abi/erc20.json";
 import POOL_REGISTRY_ABI from "./abi/poolRegistry.json";
@@ -27,13 +30,14 @@ import VTOKEN_ABI from "./abi/vToken.json";
 
 const { ethereum } = NETWORK_ADDRESSES;
 
-const BLOCKS_PER_YEAR = BigNumber.from("2628000"); // assuming a block is mined every 15 seconds
+const BLOCKS_PER_YEAR = BigNumber.from(2628000);
+const ONE_YEAR = 365 * 24 * 3600;
 
-type VTokenSymbol = "vPT-USDe-27MAR2025_Ethena" | "vPT-sUSDE-27MAR2025_Ethena";
+type VTokenSymbol = "vsUSDe_Ethena" | "vUSDC_Ethena";
 
 const vTokens: { [key in VTokenSymbol]: string } = {
-  "vPT-USDe-27MAR2025_Ethena": VPT_USDe_27MAR2025_ETHENA,
-  "vPT-sUSDE-27MAR2025_Ethena": VPT_sUSDE_27MAR2025_ETHENA,
+  vsUSDe_Ethena: VsUSDe_Ethena,
+  vUSDC_Ethena: VUSDC_Ethena,
 };
 
 interface VTokenState {
@@ -46,20 +50,20 @@ interface VTokenState {
 }
 
 const vTokenState: { [key in VTokenSymbol]: VTokenState } = {
-  "vPT-USDe-27MAR2025_Ethena": {
-    name: "Venus PT-USDe-27MAR2025 (Ethena)",
-    symbol: "vPT-USDe-27MAR2025_Ethena",
+  vsUSDe_Ethena: {
+    name: "Venus sUSDe (Ethena)",
+    symbol: "vsUSDe_Ethena",
     decimals: 8,
-    underlying: PT_USDe_27MAR2025,
+    underlying: sUSDe,
     exchangeRate: parseUnits("1", 28),
     comptroller: COMPTROLLER_ETHENA,
   },
-  "vPT-sUSDE-27MAR2025_Ethena": {
-    name: "Venus PT-sUSDE-27MAR2025 (Ethena)",
-    symbol: "vPT-sUSDE-27MAR2025_Ethena",
+  vUSDC_Ethena: {
+    name: "Venus USDC (Ethena)",
+    symbol: "vUSDC_Ethena",
     decimals: 8,
-    underlying: PT_sUSDE_27MAR2025,
-    exchangeRate: parseUnits("1", 28),
+    underlying: USDC,
+    exchangeRate: parseUnits("1", 16),
     comptroller: COMPTROLLER_ETHENA,
   },
 };
@@ -77,27 +81,27 @@ interface RiskParameters {
 }
 
 const riskParameters: { [key in VTokenSymbol]: RiskParameters } = {
-  "vPT-USDe-27MAR2025_Ethena": {
+  vsUSDe_Ethena: {
     borrowCap: "0",
-    supplyCap: "850000",
-    collateralFactor: "0.86",
-    liquidationThreshold: "0.88",
+    supplyCap: "50000000",
+    collateralFactor: "0.9",
+    liquidationThreshold: "0.92",
     reserveFactor: "0",
-    initialSupply: "10619.58410473",
+    initialSupply: "10000",
     vTokenReceiver: VTOKEN_RECEIVER,
-    protocolSeizeShareMantissa: "0.004",
-    price: BigNumber.from("944429637438332035"),
+    protocolSeizeShareMantissa: "0.010",
+    price: BigNumber.from("1139943330851019706"),
   },
-  "vPT-sUSDE-27MAR2025_Ethena": {
-    borrowCap: "0",
-    supplyCap: "12000000",
-    collateralFactor: "0.85",
-    liquidationThreshold: "0.87",
-    reserveFactor: "0",
-    initialSupply: "10653.07272377",
+  vUSDC_Ethena: {
+    borrowCap: "46000000",
+    supplyCap: "50000000",
+    collateralFactor: "0",
+    liquidationThreshold: "0",
+    reserveFactor: "0.1",
+    initialSupply: "10000",
     vTokenReceiver: VTOKEN_RECEIVER,
-    protocolSeizeShareMantissa: "0.004",
-    price: BigNumber.from("951397618866433430"),
+    protocolSeizeShareMantissa: "0.020",
+    price: BigNumber.from("1000010310000000000000000000000"),
   },
 };
 
@@ -111,20 +115,20 @@ interface InterestRateModelSpec {
 
 const interestRateModels: InterestRateModelSpec[] = [
   {
-    vTokens: ["vPT-USDe-27MAR2025_Ethena", "vPT-sUSDE-27MAR2025_Ethena"],
-    kink: "0.8",
+    vTokens: ["vUSDC_Ethena"],
+    kink: "0.92",
     base: "0",
-    multiplier: "0.08",
-    jump: "0.8",
+    multiplier: "0.16304",
+    jump: "2.5",
   },
 ];
 
 const interestRateModelAddresses: { [key in VTokenSymbol]: string } = {
-  "vPT-USDe-27MAR2025_Ethena": "",
-  "vPT-sUSDE-27MAR2025_Ethena": "",
+  vsUSDe_Ethena: "",
+  vUSDC_Ethena: "",
 };
 
-forking(21442115, async () => {
+forking(21488201, async () => {
   let poolRegistry: Contract;
 
   before(async () => {
@@ -138,7 +142,8 @@ forking(21442115, async () => {
   });
 
   await pretendExecutingVip(await vip072());
-  testForkedNetworkVipCommands("Ethena pool", await vip407());
+  testForkedNetworkVipCommands("Ethena pool", await vip411(ONE_YEAR));
+  testForkedNetworkVipCommands("Ethena pool", await vip412());
 
   describe("Post-Execution state", () => {
     before(async () => {
@@ -166,11 +171,10 @@ forking(21442115, async () => {
       it("should register Ethena pool vTokens in Ethena pool Comptroller", async () => {
         const comptroller = await ethers.getContractAt(COMPTROLLER_ABI, COMPTROLLER_ETHENA);
         const poolVTokens = await comptroller.getAllMarkets();
-        expect(poolVTokens).to.have.lengthOf(2);
-        expect(poolVTokens).to.include(vTokens["vPT-USDe-27MAR2025_Ethena"]);
-        expect(poolVTokens).to.include(vTokens["vPT-sUSDE-27MAR2025_Ethena"]);
+        expect(poolVTokens).to.have.lengthOf(4);
+        expect(poolVTokens).to.include(vTokens.vUSDC_Ethena);
+        expect(poolVTokens).to.include(vTokens.vsUSDe_Ethena);
       });
-
       for (const [symbol, { underlying }] of Object.entries(vTokenState) as [VTokenSymbol, VTokenState][]) {
         it(`should register ${symbol} in PoolRegistry`, async () => {
           const registeredVToken = await poolRegistry.getVTokenForAsset(vTokenState[symbol].comptroller, underlying);
@@ -278,17 +282,14 @@ forking(21442115, async () => {
             expect(await comptroller.minLiquidatableCollateral()).to.equal(parseUnits("100", 18));
           });
 
-          it("should have owner = NORMAL TIMELOCK", async () => {
+          it("should have owner = GUARDIAN", async () => {
             expect(await comptroller.owner()).to.equal(ethereum.NORMAL_TIMELOCK);
-          });
-
-          it("should have prime token set", async () => {
-            expect(await comptroller.prime()).to.equal(PRIME);
           });
         });
       };
 
       checkComptroller(COMPTROLLER_ETHENA, "Ethena");
+      checkIsolatedPoolsComptrollers();
     });
 
     it("Interest rates", async () => {
@@ -311,6 +312,12 @@ forking(21442115, async () => {
     it("Check Price", async () => {
       const resilientOracle = new ethers.Contract(ethereum.RESILIENT_ORACLE, RESILIENT_ORACLE_ABI, ethers.provider);
 
+      await setMaxStalePeriodInChainlinkOracle(
+        ethereum.CHAINLINK_ORACLE,
+        USDC,
+        ethers.constants.AddressZero,
+        ethereum.NORMAL_TIMELOCK,
+      );
       for (const [symbol, params] of Object.entries(riskParameters) as [VTokenSymbol, RiskParameters][]) {
         expect(await resilientOracle.getPrice(vTokenState[symbol].underlying)).equals(params.price);
         expect(await resilientOracle.getUnderlyingPrice(vTokens[symbol])).equals(params.price);
@@ -319,8 +326,22 @@ forking(21442115, async () => {
 
     it("should pause actions", async () => {
       const comptroller = await ethers.getContractAt(COMPTROLLER_ABI, COMPTROLLER_ETHENA);
-      expect(await comptroller.actionPaused(VPT_USDe_27MAR2025_ETHENA, 2)).to.be.true;
-      expect(await comptroller.actionPaused(VPT_sUSDE_27MAR2025_ETHENA, 2)).to.be.true;
+
+      expect(await comptroller.actionPaused(VsUSDe_Ethena, 2)).to.be.true;
+    });
+
+    describe("Converters", () => {
+      for (const [converterAddress, baseAsset] of Object.entries(converterBaseAssets)) {
+        const converterContract = new ethers.Contract(converterAddress, SINGLE_TOKEN_CONVERTER_ABI, ethers.provider);
+        for (const asset of underlyingAddress) {
+          it(`should set ${CONVERSION_INCENTIVE} as incentive in converter ${converterAddress}, for asset ${asset}`, async () => {
+            if (baseAsset != asset) {
+              const result = await converterContract.conversionConfigurations(baseAsset, asset);
+              expect(result.incentive).to.equal(CONVERSION_INCENTIVE);
+            }
+          });
+        }
+      }
     });
   });
 });
